Remove -Wno-sign-compare option. Fix signed/unsigned build warnings.

+++ Tetris/include/Tetris/NodeCalculatorImpl.h
@@ -166,7 +166,7 @@ protected:
         inline bool finished() const
         {
             Futile::ScopedLock lock(mMutex);
-            return mInfos.size() == mMaxDepth && mInfos.back().finished();
+            return static_cast<int>(mInfos.size()) == mMaxDepth && mInfos.back().finished();
         }
 
         inline void setFinished()
